Rename to Dyfrig and update Environment implementation.
Dyfrig was a missionary to the Welsh, and since Owen (OWIN) is a Welsh
name, and this project is essentially attempting to convert F# devs to the
OWIN platform.

The updates to Environment eliminate a lot of boxing and unboxing and
provide safer access to various values in the environment.